Download a file
The code still does basicly nothing, but now I can download a single
file. I'm not sure if I'm doing it the right way but it seems ok.

I still need to work on making it object oriented and dealing with
access tokens when they expire but my next goal is to see if I can
upload a file.